matrix_solver: Determine prodded amount using crafts per second instead of per timescale
In case a machine's speed would have it operate faster than one craft
per game-tick (1/60 second), the game would reduce this to one craft per
tick, with a different productivity value to error-correct. This
error-correction is taken into account by
`calculation.util.determine_prodded_amount`, which expects the amount of
crafts per second as its argument.

When we pass the total amount of crafts per timescale
(second/minute/hour), we can easily exceed 60 and kick the
error-correction code into effect. This would give wrong results for
minute/hour timescale, since the game is not going to error-correct in
the case of, say, 61 crafts per _minute_.

To address this, pass the amount of crafts per second, without
considering the timescale.