Faster allowed free items search
Approximate proof:

Adding a linearly independent column is equal to adding a linearly
independent row to a transposed matrix. We want to add a row $x$ with
exactly one 1 at position $i$. Lets analyze the matrix after performing
reduction on all rows except our row $x$.

Forward:
If row is linearly independent => $i$ is in linearly dependent column.

Suppose it is not in a linearly dependent column. Then it is in a
column with a pivot element in row $y$. Because it is in a linearly
independent column, there are no other non-zero elements in row $y$.
So row $y$ matches our row $x$ exactly => our row is not
linearly independent.

Backward:
If $i$ is in a linearly dependent column => row is linearly independent

a) It is in a column without a pivot element => it becomes the
pivot element.
b) It is in a column with a pivot element in row $y$. Then row $y$ will
have at least one more non-zero element. So after subtracting row $y$
from our row $x$, the non-zero element in the row $x$ will be in a
different position. Repeat until it is in a column without a pivot
element. Note that each time the element moves to the right, so it
cannot go on forever.