Simplify beacon UI when beacon overload is active.
Space Exploration (and some other other mods that offer the same feature in
isolation) rebalances beacons, by only allowing machines to be covered by at
most one beacon at a time (machines under the influence of two or more beacons
are disabled instead).

In this context, the "amount" field in the beacon selection UI is not
necessary, as the only legal value for it is 1. This change cleans up the UI
and forces the beacon amount to be always 1.

(This change keeps the "total" field, as that is still relevant for power
consumption calculations.)