Include burnt results in lines, to support Py ash and canisters.
A line's products are the recipe products plus the burnt result of any
fuel. This reveals the burnt result created by any line you already
have, so that you know to handle it. In the matrix solver you can
process the byproduct as usual (e.g. refilling fluid canisters).

What burnt result will be produced depends on the recipe, the machine,
and the fuel. This does not any recipes which fill demand for the
burnt result (e.g. you need ash, we don't suggest burning coal to make
steam in a boiler). This is pretty wonky, but it's an improvement from
not having ash show up at all. You can fake it by adding a one steam
as an output from your factory, and the extra steam will turn into
a byproduct. Or you can paste the ash-producing line you want in from
another factory and the solver will use it.